# task-managament-api/src/main/resources/task-management-api-definition.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
paths:
...
  /api/v1/tasks:
    post:
        tags:
          - tasks
        description: "Create a new task for the authenticated user"
        operationId: createTask
        requestBody:
            content:
                application/json:
                    schema:
                      $ref: "#/components/schemas/TaskRequest"
        responses:
          '201':
            description: "Task created successfully"
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/TaskResponse"
          '400':
            description: "Invalid input data"
          '401':
            description: "Unauthorized access - invalid or missing JWT"
...
components:
  schemas:
...
    TaskRequest:
        type: object
        required:
            - title
            - projectId
            - taskPriority
        properties:
            title:
                type: string
            description:
                type: string
            projectId:
                type: string
                format: uuid
            sprintId:
                type: string
                format: uuid
            assigneeId:
                type: string
                format: uuid
            taskPriority:
              type: string
              enum: [LOW, MEDIUM, HIGH]
    TaskResponse:
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        taskStatus:
          type: string
        taskPriority:
          type: string
        projectId:
          type: string
          format: uuid
        assigneeId:
          type: string
          format: uuid
...